+++ R/venn_data.R
@@ -1,0 +1,26 @@
+#' @title Venn data with multiple gene sets.
+#' @description Venn data with multiple gene sets.
+#' @author benben-miao
+#'
+#' @docType data
+#' @format Dataframe: include multiple gene sets.
+#' @usage data(venn_data)
+#'
+#' @keywords datasets
+#'
+#' @references https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/VennPlot/
+#'
+#' @examples
+#' # 1. Library TOmicsVis package
+#' library(TOmicsVis)
+#'
+#' # 2. Load example data
+#' data(venn_data)
+#'
+#' # 3. View example data
+#' venn_data
+#'
+"venn_data"
+
+# venn_data <- data
+# usethis::use_data(venn_data)